Fix terminal error when calling clasp logs with invalid projectId and refactor clasp logs (#535)
* Add error for projectid error

* Validate projectId and fix #518

* remove unused import

* Add project id validation method

* Move switch statement in a method

* linting

* Refactor clasp logs

* Refactor clasp logs

* Update index.ts

* Update index.ts
